Split DeleteCommand into DeleteTaskCommand, DeleteModuleCommand
